# docs/swagger.yaml
definitions:
  api.AddMemberRequestBody:
    properties:
      executor_id:
        type: string
      group_chat_id:
        type: string
      role:
        type: string
      user_account_id:
        type: string
    type: object
  api.AddMemberResponseSuccessBody:
    properties:
      group_chat_id:
        type: string
    type: object
  api.CreateGroupChatRequestBody:
    properties:
      executor_id:
        example: UserAccount-01H42K4ABWQ5V2XQEP3A48VE0Z
        type: string
      name:
        example: group-chat-name-1
        type: string
    type: object
  api.CreateGroupChatResponseSuccessBody:
    properties:
      group_chat_id:
        example: GroupChat-01H42K4ABWQ5V2XQEP3A48VE0Z
        type: string
    type: object
  api.DeleteGroupChatRequestBody:
    properties:
      executor_id:
        type: string
      group_chat_id:
        type: string
    type: object
  api.DeleteGroupChatResponseSuccessBody:
    properties:
      group_chat_id:
        type: string
    type: object
  api.DeleteMessageRequestBody:
    properties:
      executor_id:
        type: string
      group_chat_id:
        type: string
      message_id:
        type: string
      user_account_id:
        type: string
    type: object
  api.DeleteMessageResponseSuccessBody:
    properties:
      group_chat_id:
        type: string
    type: object
  api.GroupChatResponseErrorBody:
    properties:
      message:
        type: string
    type: object
  api.PostMessageRequestBody:
    properties:
      executor_id:
        type: string
      group_chat_id:
        type: string
      message:
        type: string
      user_account_id:
        type: string
    type: object
  api.PostMessageResponseSuccessBody:
    properties:
      group_chat_id:
        type: string
      message_id:
        type: string
    type: object
  api.RenameGroupChatRequestBody:
    properties:
      executor_id:
        type: string
      group_chat_id:
        type: string
      name:
        type: string
    type: object
  api.RenameGroupChatResponseSuccessBody:
    properties:
      group_chat_id:
        type: string
    type: object
info:
  contact: {}
  description: This is Write API Server for GroupChat.
  license:
    name: Junichi Kato
  title: GroupChat Write API
  version: "1.0"
paths:
  /v1/:
    get:
      description: Index
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Index
  /v1/group-chat/add-member:
    post:
      description: Add a member to a group chat
      parameters:
      - description: AddMemberRequestBody
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.AddMemberRequestBody'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.AddMemberResponseSuccessBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.GroupChatResponseErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.GroupChatResponseErrorBody'
      summary: Add a member to a group chat
  /v1/group-chat/create:
    post:
      description: Create a group chat
      parameters:
      - description: CreateGroupChatRequestBody
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.CreateGroupChatRequestBody'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.CreateGroupChatResponseSuccessBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.GroupChatResponseErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.GroupChatResponseErrorBody'
      summary: Create a group chat
  /v1/group-chat/delete:
    post:
      description: Delete a group chat
      parameters:
      - description: DeleteGroupChatRequestBody
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.DeleteGroupChatRequestBody'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.DeleteGroupChatResponseSuccessBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.GroupChatResponseErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.GroupChatResponseErrorBody'
      summary: Delete a group chat
  /v1/group-chat/delete-message:
    post:
      description: Delete a message from a group chat
      parameters:
      - description: DeleteMessageRequestBody
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.DeleteMessageRequestBody'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.DeleteMessageResponseSuccessBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.GroupChatResponseErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.GroupChatResponseErrorBody'
      summary: Delete a message from a group chat
  /v1/group-chat/post-message:
    post:
      description: Post a message to a group chat
      parameters:
      - description: PostMessageRequestBody
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.PostMessageRequestBody'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.PostMessageResponseSuccessBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.GroupChatResponseErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.GroupChatResponseErrorBody'
      summary: Post a message to a group chat
  /v1/group-chat/rename:
    post:
      description: Rename a group chat
      parameters:
      - description: RenameGroupChatRequestBody
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.RenameGroupChatRequestBody'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.RenameGroupChatResponseSuccessBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.GroupChatResponseErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.GroupChatResponseErrorBody'
      summary: Rename a group chat
swagger: "2.0"